Add layer property reducer
